remove include of non existing ext/PMFRGMPIExt/Solver.jl
this was causing mysterious failure when running the test suite.
the actual error was not visible then,
possibly because it was masked
by the assertion failure triggered when the extension is not loaded
(see runtests.jl).

When doing

using MPI, PencilArrays
using PMFRG

then the extension compilation was triggered,
but then the error is visible
(contrarily to during the test)